mod interface

+++ rosen/modules/effect/color_picker/include/color_extract.h
@@ -73,7 +73,7 @@ public:
 
 protected:
     ColorExtract(std::shared_ptr<Media::PixelMap> pixmap);
-    ColorExtract(std::shared_ptr<Media::PixelMap> pixmap, float* coordinates);
+    ColorExtract(std::shared_ptr<Media::PixelMap> pixmap, double* coordinates);
 
 private:
